backend/requirements.txt
fastapi
joblib
scikit-learn==1.2.2
numpy==1.26.4
python-dotenv
netCDF4
earthaccess
pyproj==3.7.1
opencv-python-headless
uvicorn[standard]
pydantic
tensorflow
xgboost
Pillow
ultralytics
python-multipart
pybase64